+++ internal/config/config.go
@@ -61,26 +61,23 @@ type Prometheus struct {
 // Load initializes the Config variable from file and the environment.
 //  Environment variables take precedence over the file.
 func Load(file io.Reader) error {
-	var fileErr error
 	Config = config{}
 
-	if file != nil {
-		if _, err := toml.DecodeReader(file, &Config); err != nil {
-			fileErr = fmt.Errorf("decode config: %v", err)
-		}
+	if _, err := toml.DecodeReader(file, &Config); err != nil {
+		return fmt.Errorf("load toml: %v", err)
 	}
 
-	err := envconfig.Process("gitaly", &Config)
-	if err != nil {
-		log.WithError(err).Fatal("process environment variables")
+	if err := envconfig.Process("gitaly", &Config); err != nil {
+		return fmt.Errorf("envconfig: %v", err)
 	}
 
-	return fileErr
+	return nil
 }
 
 // Validate checks the current Config for sanity.
 func Validate() error {
 	for _, err := range []error{
+		validateListeners(),
 		validateStorages(),
 		validateToken(),
 		SetGitPath(),
@@ -90,6 +87,13 @@ func Validate() error {
 		if err != nil {
 			return err
 		}
+	}
+	return nil
+}
+
+func validateListeners() error {
+	if len(Config.SocketPath) == 0 && len(Config.ListenAddr) == 0 {
+		return fmt.Errorf("invalid listener config: at least one of socket_path and listen_addr must be set")
 	}
 	return nil
 }
@@ -127,22 +131,22 @@ func validateIsDirectory(path, name string) error {
 
 func validateStorages() error {
 	if len(Config.Storages) == 0 {
-		return fmt.Errorf("config: no storage configurations found. Is your gitaly.config correctly configured? https://gitlab.com/gitlab-org/gitaly/issues/397")
+		return fmt.Errorf("no storage configurations found. Are you using the right format? https://gitlab.com/gitlab-org/gitaly/issues/397")
 	}
 
 	seenNames := make(map[string]bool)
 	for _, st := range Config.Storages {
 		if st.Name == "" {
-			return fmt.Errorf("config: empty storage name in %v", st)
+			return fmt.Errorf("empty storage name in %v", st)
 		}
 
 		if st.Path == "" {
-			return fmt.Errorf("config: empty storage path in %v", st)
+			return fmt.Errorf("empty storage path in %v", st)
 		}
 
 		name := st.Name
 		if seenNames[name] {
-			return fmt.Errorf("config: storage %q is defined more than once", name)
+			return fmt.Errorf("storage %q is defined more than once", name)
 		}
 		seenNames[name] = true
 	}