+++ internal/service/commit/find_commits.go
@@ -47,6 +47,8 @@ func (s *server) FindCommits(req *pb.FindCommitsRequest, stream pb.CommitService
 	return rubyserver.Proxy(func() error {
 		resp, err := rubyStream.Recv()
 		if err != nil {
+			md := rubyStream.Trailer()
+			stream.SetTrailer(md)
 			return err
 		}
 		return stream.Send(resp)