+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 
 - Fix path bug in CommitService::FindCommits
   https://gitlab.com/gitlab-org/gitaly/merge_requests/364
+- Fail harder during startup, fix version string
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/379
 
 v0.42.0